Set charging_demand and consumption to 0 where timerange is exceeded #66
Introduced in 97076f507cd46b41a83286a41c78b9640c725852 the final timestep is set to upper bound of when exceeded. This lead to wrong values for charging_demand and consumption. In this fix, both are set to 0 and for charging events soc_end is set to soc_start. Note: this leaves the consumption being wrong!